lect7-demo/requirements.txt
requests
python-dotenv
flask